+++ Web/Scotty/Internal/Types.hs
@@ -5,27 +5,31 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# LANGUAGE RecordWildCards #-}
+
 module Web.Scotty.Internal.Types where
 
 import           Blaze.ByteString.Builder (Builder)
 
 import           Control.Applicative
+import           Control.Exception (Exception)
 import qualified Control.Exception as E
+import qualified Control.Monad as Monad
+import           Control.Monad (MonadPlus(..))
 import           Control.Monad.Base (MonadBase, liftBase, liftBaseDefault)
 import           Control.Monad.Catch (MonadCatch, catch, MonadThrow, throwM)
 import           Control.Monad.Error.Class
 import qualified Control.Monad.Fail as Fail
-import           Control.Monad.Reader
-import           Control.Monad.State
+import           Control.Monad.IO.Class (MonadIO)
+import           Control.Monad.Reader (MonadReader(..), ReaderT, mapReaderT)
+import           Control.Monad.State.Strict (MonadState(..), State, StateT, mapStateT)
+import           Control.Monad.Trans.Class (MonadTrans(..))
 import           Control.Monad.Trans.Control (MonadBaseControl, StM, liftBaseWith, restoreM, ComposeSt, defaultLiftBaseWith, defaultRestoreM, MonadTransControl, StT, liftWith, restoreT)
 import           Control.Monad.Trans.Except
 
 import qualified Data.ByteString as BS
 import           Data.ByteString.Lazy.Char8 (ByteString)
 import           Data.Default.Class (Default, def)
-#if !(MIN_VERSION_base(4,8,0))
-import           Data.Monoid (mempty)
-#endif
 import           Data.String (IsString(..))
 import           Data.Text.Lazy (Text, pack)
 import           Data.Typeable (Typeable)
@@ -36,6 +40,9 @@ import           Network.Wai hiding (Middleware, Application)
 import qualified Network.Wai as Wai
 import           Network.Wai.Handler.Warp (Settings, defaultSettings)
 import           Network.Wai.Parse (FileInfo)
+
+import           Prelude ()
+import           Prelude.Compat
 
 --------------------- Options -----------------------
 data Options = Options { verbose :: Int -- ^ 0 = silent, 1(def) = startup banner
@@ -51,6 +58,13 @@ data Options = Options { verbose :: Int -- ^ 0 = silent, 1(def) = startup banner
 instance Default Options where
     def = Options 1 defaultSettings
 
+newtype RouteOptions = RouteOptions { maxRequestBodySize :: Maybe Kilobytes -- max allowed request size in KB
+                                    }
+
+instance Default RouteOptions where
+    def = RouteOptions Nothing
+
+type Kilobytes = Int
 ----- Transformer Aware Applications/Middleware -----
 type Middleware m = Application m -> Application m
 type Application m = Request -> m Response
@@ -60,10 +74,11 @@ data ScottyState e m =
     ScottyState { middlewares :: [Wai.Middleware]
                 , routes :: [Middleware m]
                 , handler :: ErrorHandler e m
+                , routeOptions :: RouteOptions
                 }
 
 instance Default (ScottyState e m) where
-    def = ScottyState [] [] Nothing
+    def = ScottyState [] [] Nothing def
 
 addMiddleware :: Wai.Middleware -> ScottyState e m -> ScottyState e m
 addMiddleware m s@(ScottyState {middlewares = ms}) = s { middlewares = m:ms }
@@ -74,15 +89,21 @@ addRoute r s@(ScottyState {routes = rs}) = s { routes = r:rs }
 addHandler :: ErrorHandler e m -> ScottyState e m -> ScottyState e m
 addHandler h s = s { handler = h }
 
+updateMaxRequestBodySize :: RouteOptions -> ScottyState e m -> ScottyState e m
+updateMaxRequestBodySize RouteOptions { .. } s@ScottyState { routeOptions = ro } =
+    let ro' = ro { maxRequestBodySize = maxRequestBodySize }
+    in s { routeOptions = ro' }
+
 newtype ScottyT e m a = ScottyT { runS :: State (ScottyState e m) a }
     deriving ( Functor, Applicative, Monad )
 
 
 ------------------ Scotty Errors --------------------
-data ActionError e = Redirect Text
-                   | Next
-                   | Finish
-                   | ActionError e
+data ActionError e
+  = Redirect Text
+  | Next
+  | Finish
+  | ActionError Status e
 
 -- | In order to use a custom exception type (aside from 'Text'), you must
 -- define an instance of 'ScottyError' for that type.
@@ -95,13 +116,17 @@ instance ScottyError Text where
     showError = id
 
 instance ScottyError e => ScottyError (ActionError e) where
-    stringError = ActionError . stringError
+    stringError = ActionError status500 . stringError
     showError (Redirect url)  = url
     showError Next            = pack "Next"
     showError Finish          = pack "Finish"
-    showError (ActionError e) = showError e
+    showError (ActionError _ e) = showError e
 
 type ErrorHandler e m = Maybe (e -> ActionT e m ())
+
+data ScottyException = RequestException BS.ByteString Status deriving (Show, Typeable)
+
+instance Exception ScottyException
 
 ------------------ Scotty Actions -------------------
 type Param = (Text, Text)
@@ -139,7 +164,7 @@ instance Default ScottyResponse where
 newtype ActionT e m a = ActionT { runAM :: ExceptT (ActionError e) (ReaderT ActionEnv (StateT ScottyResponse m)) a }
     deriving ( Functor, Applicative, MonadIO )
 
-instance (Monad m, ScottyError e) => Monad (ActionT e m) where
+instance (Monad m, ScottyError e) => Monad.Monad (ActionT e m) where
     return = ActionT . return
     ActionT m >>= k = ActionT (m >>= runAM . k)
 #if !(MIN_VERSION_base(4,13,0))
@@ -165,7 +190,7 @@ instance (Monad m, ScottyError e) => MonadPlus (ActionT e m) where
             Left  _ -> runExceptT n
             Right r -> return $ Right r
 
-instance MonadTrans (ActionT e) where
+instance ScottyError e => MonadTrans (ActionT e) where
     lift = ActionT . lift . lift . lift
 
 instance (ScottyError e, Monad m) => MonadError (ActionError e) (ActionT e m) where
@@ -184,7 +209,7 @@ instance (MonadThrow m, ScottyError e) => MonadThrow (ActionT e m) where
 instance (MonadCatch m, ScottyError e) => MonadCatch (ActionT e m) where
     catch (ActionT m) f = ActionT (m `catch` (runAM . f))
 
-instance MonadTransControl (ActionT e) where
+instance ScottyError e => MonadTransControl (ActionT e) where
      type StT (ActionT e) a = StT (StateT ScottyResponse) (StT (ReaderT ActionEnv) (StT (ExceptT (ActionError e)) a))
      liftWith = \f ->
         ActionT $  liftWith $ \run  ->
@@ -197,6 +222,58 @@ instance (ScottyError e, MonadBaseControl b m) => MonadBaseControl b (ActionT e 
     type StM (ActionT e m) a = ComposeSt (ActionT e) m a
     liftBaseWith = defaultLiftBaseWith
     restoreM     = defaultRestoreM
+
+instance (MonadReader r m, ScottyError e) => MonadReader r (ActionT e m) where
+    {-# INLINE ask #-}
+    ask = lift ask
+    {-# INLINE local #-}
+    local f = ActionT . mapExceptT (mapReaderT (mapStateT $ local f)) . runAM
+
+instance (MonadState s m, ScottyError e) => MonadState s (ActionT e m) where
+    {-# INLINE get #-}
+    get = lift get
+    {-# INLINE put #-}
+    put = lift . put
+
+instance (Semigroup a) => Semigroup (ScottyT e m a) where
+  x <> y = (<>) <$> x <*> y
+
+instance
+  ( Monoid a
+#if !(MIN_VERSION_base(4,11,0))
+  , Semigroup a
+#endif
+#if !(MIN_VERSION_base(4,8,0))
+  , Functor m
+#endif
+  ) => Monoid (ScottyT e m a) where
+  mempty = return mempty
+#if !(MIN_VERSION_base(4,11,0))
+  mappend = (<>)
+#endif
+
+instance
+  ( Monad m
+#if !(MIN_VERSION_base(4,8,0))
+  , Functor m
+#endif
+  , Semigroup a
+  ) => Semigroup (ActionT e m a) where
+  x <> y = (<>) <$> x <*> y
+
+instance
+  ( Monad m, ScottyError e, Monoid a
+#if !(MIN_VERSION_base(4,11,0))
+  , Semigroup a
+#endif
+#if !(MIN_VERSION_base(4,8,0))
+  , Functor m
+#endif
+  ) => Monoid (ActionT e m a) where
+  mempty = return mempty
+#if !(MIN_VERSION_base(4,11,0))
+  mappend = (<>)
+#endif
 
 ------------------ Scotty Routes --------------------
 data RoutePattern = Capture   Text